\name{dce.spline}
\alias{dce.spline}
\title{Bayesian P-Splines for Modeling Dynamic Contrasat-Enhanced MRI Data}}
\description{
  Quantitative analysis of DCE-MRI typically involves the convolution of
  an arterial input function (AIF) with a nonlinear pharmacokinetic
  model of the contrast agent concentration.  This function takes a
  semi-parametric penalized spline smoothing approach, with which the
  AIF is convolved with a set of B-splines to produce a design matrix
  using locally adaptive smoothing parameters based on Bayesian
  penalized spline models (P-splines).} 
\usage{
dce.spline(data, time, input, ...)
}
\arguments{
}
\value{
}
\details{
}
\references{
}
\seealso{
}
\examples{
}
\author{Volker Schmid}
\keyword{}
