\name{Seed Growing for a 4D Array}
\alias{extract.aif}
\title{Seed Growing for a 4D Array}
\description{
  Seed growing algorithm to find voxels in a three-dimensional array
  according to their correlation to a seed voxel.  The corrlation is
  measured according to a fourth dimension in the array.
}
\usage{
extract.aif(img, x, y, z, thresh=0.9)
}
\arguments{
  \item{img}{}
  \item{x,y,z}{}
  \item{thresh}{}
}
\value{
}
\details{
}
%\references{}
%\seealso{}
%\examples{}
\author{Volker Schmid}
\keyword{}
