\name{hotmetal}
\alias{hotmetal}
\title{Hot Metal Color Table}
\description{
  The hotmetal color table patterned after that used in Matlab. 
}
\usage{
hotmetal(n=64)
}
\arguments{
  \item{n}{Number of color levels; 64 is the default.}
}
\details{
  See \code{tim.colors} in the \code{fields} package.
}
\value{
  A vector of character strings giving the colors in hexadecimal
  format.
}
\seealso{
  \code{\link{topo.colors}},
  \code{\link{terrain.colors}}}
%\examples{}
\keyword{aplot}
