% Part of the dcemri package for R
% Distributed under the BSD license: see dcemri/COPYING
%
% $Id: integerTranslation.Rd 191 2009-08-25 15:12:31Z bjw34032 $

\name{integerTranslation}
\alias{integerTranslation}
\alias{invertIntegerTranslation}
\title{integerTranslation}
\description{
  ...
}
\usage{
integerTranslation(nim, data, verbose=FALSE)
invertIntegerTranslation(nim, verbose=FALSE)
}
\arguments{
  \item{nim}{}
  \item{data}{}
  \item{verbose}{}
}
\details{
  ...
}
\value{
  ...
}
%\references{}
%\seealso{}
%\examples{}
\author{Andrew Thornton <\email{zeripath@users.sourceforge.net}>}
%\keyword{}
