% Part of the dcemri package for R
% Distributed under the BSD license: see dcemri/COPYING
%
% $Id: performPermutation.Rd 191 2009-08-25 15:12:31Z bjw34032 $

\name{performPermutation}
\alias{performPermutation}
\title{performPermutation}
\description{
  ...
}
\usage{
performPermutation(T, real.dimensions, data, verbose=FALSE)
}
\arguments{
  \item{T}{}
  \item{real.dimensions}{}
  \item{data}{}
  \item{verbose}{}
}
\details{
  ...
}
\value{
  ...
}
%\references{}
%\seealso{}
%\examples{}
\author{Andrew Thornton <\email{zeripath@users.sourceforge.net}>}
%\keyword{}
