% Part of the dcemri package for R
% Distributed under the BSD license: see dcemri/COPYING
%
% $Id: reorient.Rd 191 2009-08-25 15:12:31Z bjw34032 $

\name{reorient}
\alias{reorient}
\alias{inverseReorient}
\title{reorient}
\description{
  ...
}
\usage{
reorient(nim, data, verbose=FALSE, invert=FALSE)
inverseReorient(nim, verbose=FALSE)
}
\arguments{
  \item{nim}{}
  \item{data}{}
  \item{verbose}{}
  \item{invert}{}
}
\details{
  ...
}
\value{
  ...
}
%\references{}
%\seealso{}
%\examples{}
\author{Andrew Thornton <\email{zeripath@users.sourceforge.net}>}
%\keyword{}
