% Part of the dcemri package for R
% Distributed under the BSD license: see dcemri/COPYING
%
% $Id: transformCoordinate.Rd 191 2009-08-25 15:12:31Z bjw34032 $

\name{translateCoordinate}
\alias{translateCoordinate}
\title{translateCoordinate}
\description{
  ...
}
\usage{
translateCoordinate(i, nim, verbose=FALSE)
}
\arguments{
  \item{i}{}
  \item{nim}{}
  \item{verbose}{}
}
\details{
  ...
}
\value{
  ...
}
%\references{}
%\seealso{}
%\examples{}
\author{Andrew Thornton <\email{zeripath@users.sourceforge.net}>}
%\keyword{}
