% Part of the dcemri package for R
% Distributed under the BSD license: see dcemri/COPYING
%
% $Id: write_anlz.Rd 191 2009-08-25 15:12:31Z bjw34032 $

\name{writeANALYZE}
\alias{writeANALYZE}
\title{writeANALYZE}
\description{
  ...
}
\usage{
}
\arguments{
}
\details{
}
\value{
}
\references{
}
\seealso{
  \code{\link{writeNIfTI}}
}
%\examples{}
\author{}
%\keyword{}
