\name{Arterial Input Functions}
\alias{aif.orton.exp}
\alias{orton.exp.lm}
\title{Arterial Input Functions}
\description{
}
\usage{
aif.orton.exp(tt, AB, muB, AG, muG)
orton.exp.lm(tt, aif, guess=c(100, 10, 1, 0.1), nprint=0)
}
\arguments{
  \item{tt}
  \item{AG,muB,AG,muG}{}
  \item{aif}{}
  \item{guess}{}
  \item{nprint}{}
}
\value{
}
\details{
}
\references{
}
%\seealso{}
%\examples{}
\author{Brandon Whitcher}
\keyword{}
