\name{Buckley}
\alias{buckley}
\alias{breast}
\alias{meningioma}
\title{Simuluted Data from Buckley (2002)}
\description{
  Simuluted data from Buckley (2002).
}
\usage{
data("buckley")
}
\format{
  Two lists are created (breast and meningioma) that contain the
  simulated time curves and all associated kinetic parameter values.
}
\source{
  See below.
}
\references{
  Buckley, D.L. (2002) Uncertainty in the Analysis of Tracer Kinetics
  Using Dynamic Contrast-Enhanced \eqn{T_1}-weighted MRI, \emph{Magnetic
    Resonance in Medicine} \bold{47}, 601-606.
}
\keyword{datasets}
