\name{dcemri}
\alias{dcemri.ext}
\alias{dcemri.fit}
\title{Routines to Fit Pharmacokinetic Models to DCE-MRI Data}
\description{
  TBA
}
\usage{
dcemri.fit(conc, tmin, img.mask, aif, nprint=0)
dcemri.ext(conc, tmin, img.mask, aif, nprint=0)
}
\arguments{
  \item{conc}{}
  \item{tmin}{}
  \item{img.mask}{}
  \item{aif}{}
  \item{nprint}{}
}
\value{
}
\details{
}
\references{
}
\seealso{
  dcemri.nls, dcemri2.nls
}
%\examples{}
\author{Brandon Whitcher}
\keyword{}
