\name{dcemri}
\alias{dcemri.lm}
\title{Pharmacokinetic Models for Dynamic Contrast-Enhanced MRI Data}
\description{
  TBA
}
\usage{
dcemri.lm(conc, time, img.mask, model="extended", aif="tofts.kermode",
          nprint=0, user=NULL, ...)
}
\arguments{
  \item{conc}{}
  \item{time}{}
  \item{img.mask}{}
  \item{model}{}
  \item{aif}{}
  \item{nprint}{}
  \item{user}{}
  \item{...}{}
}
\value{
}
\details{
}
\references{
}
\seealso{
  dce.bayes, dce.spline
}
%\examples{}
\author{Brandon Whitcher, Volker Schmid}
\keyword{}
